+++ backend/src/main/java/com/stampcrush/backend/api/visitor/profile/VisitorProfilesCommandApiController.java
@@ -2,6 +2,7 @@ package com.stampcrush.backend.api.visitor.profile;
 
 import com.stampcrush.backend.api.visitor.profile.request.VisitorProfilesPhoneNumberUpdateRequest;
 import com.stampcrush.backend.application.visitor.profile.VisitorProfilesCommandService;
+import com.stampcrush.backend.application.visitor.profile.dto.VisitorProfilesLinkDataDto;
 import com.stampcrush.backend.config.resolver.CustomerAuth;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
